update to match lib9

+++ libkern/qsort.c
@@ -2,6 +2,7 @@
  * qsort -- simple quicksort
  */
 
+#include "lib9.h"
 typedef
 struct
 {
@@ -116,7 +117,7 @@ qsort(void *va, long n, long es, int (*cmp)(void*, void*))
 	s.cmp = cmp;
 	s.es = es;
 	s.swap = swapi;
-	if(((long)va | es) % sizeof(long))
+	if(((uintptr)va | es) % sizeof(long))
 		s.swap = swapb;
 	qsorts((char*)va, n, &s);
 }
